refactor: 重命名兼容性训练配置和训练器为兼容性预测
