+++ src/trains/configs/compatibility_prediction_train_config.py
@@ -5,12 +5,12 @@ from .base_train_config import BaseTrainConfig
 
 
 @dataclass
-class CompatibilityTrainConfig(BaseTrainConfig):
+class CompatibilityPredictionTrainConfig(BaseTrainConfig):
     find_unused_parameters = True
     batch_size: int = 512
     dataloader_workers: int = 4
     polyvore_type: Literal['nondisjoint', 'disjoint'] = 'nondisjoint'
-    name: str = 'compatibility_train'
+    name: str = 'compatibility_prediction'
     auto_save_checkpoint: bool = True
     load_image: bool = False
     learning_rate: float = 2e-5 # learning rate
